hsy商户

+++ service/bill/bill-api/src/main/java/com/jkm/hss/bill/service/ShareProfitService.java
@@ -22,7 +22,7 @@ public interface ShareProfitService {
      * @param req
      * @return
      */
-   long selectAccountListCount(JkmProfitRequest req);
+   int selectAccountListCount(JkmProfitRequest req);
 
 
 }
